Update `firestore-rules` to `2.x.x`
With the new API, it also made sense to rename `firestore-rules`
to `firebase-rules`, alongside `test/utils.ts` and `test/setup.ts`.
Added support for `tsconfig` paths, using `tsconfig-paths`.
