+++ firebase/src/scripts/dummy-script.ts
@@ -1,4 +1,4 @@
-import { dummy } from "../core/dummy";
+import { dummy } from "@core/dummy";
 
 export const scriptValue = 1;
 console.log(dummy.value);
