+++ firebase/test/firebase-rules/collections-categories.spec.ts
@@ -1,38 +1,38 @@
 import * as rulesTesting from "@firebase/rules-unit-testing";
-import * as utils from "../utils";
+import * as utils from "@test/firebase-rules/utils";
 
-describe("collections/{doc}/memos", () => {
-  const collectionId = "collections/any_id/memos";
+describe("collections_categories/", () => {
+  const collectionId = "collections_categories";
   let collectionsRef: utils.CollectionReference;
-  let firestore: utils.ClientFirestore;
+  let firestore: utils.Firestore;
 
   describe("authenticated users", () => {
-    before(async () => {
-      firestore = await utils.createMyFirestore();
+    before(() => {
+      firestore = utils.createMyFirestore();
       collectionsRef = firestore.collection(collectionId);
     });
 
-    it("should be able to read collections' memos", async () => {
+    it("should be able to read collections categories", async () => {
       await rulesTesting.assertSucceeds(collectionsRef.get());
     });
 
-    it("should be denied to write any collection memo", async () => {
+    it("should be denied to write any collection category", async () => {
       await rulesTesting.assertFails(collectionsRef.add({}));
       await rulesTesting.assertFails(collectionsRef.doc().set({ any: "any" }));
     });
   });
 
   describe("unauthenticated users", () => {
-    before(async () => {
-      firestore = await utils.createFirestore();
+    before(() => {
+      firestore = utils.createFirestore();
       collectionsRef = firestore.collection(collectionId);
     });
 
-    it("should be denied to read collections' memos", async () => {
+    it("should be denied to read collections categories", async () => {
       await rulesTesting.assertFails(collectionsRef.get());
     });
 
-    it("should be denied to write to any collection memo", async () => {
+    it("should be denied to write to any collection category", async () => {
       await rulesTesting.assertFails(collectionsRef.add({}));
       await rulesTesting.assertFails(collectionsRef.doc().set({ any: "any" }));
     });
