+++ firebase/test/firebase-rules/collections.spec.ts
@@ -1,10 +1,10 @@
 import * as rulesTesting from "@firebase/rules-unit-testing";
-import * as utils from "./utils";
+import * as utils from "@test/firebase-rules/utils";
 
-describe("collections_categories/", () => {
-  const collectionId = "collections_categories";
+describe("collections/", () => {
+  const collectionId = "collections";
   let collectionsRef: utils.CollectionReference;
-  let firestore: utils.ClientFirestore;
+  let firestore: utils.Firestore;
 
   describe("authenticated users", () => {
     before(() => {
@@ -12,11 +12,11 @@ describe("collections_categories/", () => {
       collectionsRef = firestore.collection(collectionId);
     });
 
-    it("should be able to read collections categories", async () => {
+    it("should be able to read collections", async () => {
       await rulesTesting.assertSucceeds(collectionsRef.get());
     });
 
-    it("should be denied to write any collection category", async () => {
+    it("should be denied to write any collection", async () => {
       await rulesTesting.assertFails(collectionsRef.add({}));
       await rulesTesting.assertFails(collectionsRef.doc().set({ any: "any" }));
     });
@@ -28,11 +28,11 @@ describe("collections_categories/", () => {
       collectionsRef = firestore.collection(collectionId);
     });
 
-    it("should be denied to read collections categories", async () => {
+    it("should be denied to read collections", async () => {
       await rulesTesting.assertFails(collectionsRef.get());
     });
 
-    it("should be denied to write to any collection category", async () => {
+    it("should be denied to write to any collection", async () => {
       await rulesTesting.assertFails(collectionsRef.add({}));
       await rulesTesting.assertFails(collectionsRef.doc().set({ any: "any" }));
     });
