--- firebase/test/firestore-rules/setup.ts
@@ -1,6 +1,0 @@
-import * as utils from "./utils";
-
-// Clears all firestore data after each test
-afterEach(async () => {
-  await utils.clearFirestore();
-});
